Agregue varias validaciones, Where para poder seguir despues de un Ins, Upd o Del y un Extension Method

+++ Logic/CategoriesLogic.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data.Entity;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
@@ -18,12 +19,11 @@ namespace Logic
 
         public List<Categories> GetAll()
         {
-            NorthwindContext context = new NorthwindContext();
             return context.Categories.ToList();
         }
+       
         public Categories GetOne(int id) 
         {
-            context.Categories.FirstOrDefault(r => r.CategoryID.Equals(id));
             return context.Categories.FirstOrDefault(r => r.CategoryID.Equals(id));
         }
 
@@ -44,20 +44,39 @@ namespace Logic
         public Categories Update(Categories entity)
         {
             Categories categoriaAEditar = GetOne(entity.CategoryID);
-            categoriaAEditar.CategoryName = entity.CategoryName;
-            categoriaAEditar.Description = entity.Description;
-            context.SaveChanges();
-            Categories T = null;
-            return T;
+            if (categoriaAEditar == null)
+            {
+                Console.WriteLine("El número de ID ingresado no es válido.");
+                return categoriaAEditar;
+                
+            }
+            else
+            {
+                categoriaAEditar.CategoryName = entity.CategoryName;
+                categoriaAEditar.Description = entity.Description;
+                context.SaveChanges();
+                Console.WriteLine("");
+                Console.WriteLine($"LA MODIFICACIÓN FUE EXITOSA: {System.Environment.NewLine} CATEGORIA {entity.CategoryID}: {entity.CategoryID} CON DESCRIPCIÓN: {entity.Description}. {System.Environment.NewLine}");
+                return categoriaAEditar;
+            }
         }
 
         public Categories Delete(Categories entity)
         {
             Categories categoriaAEliminar = GetOne(entity.CategoryID);
-            context.Categories.Remove(categoriaAEliminar);
-            context.SaveChanges();
-            Categories T = null;
-            return T;
+            if (categoriaAEliminar != null)
+            {
+                context.Categories.Remove(categoriaAEliminar);
+                context.SaveChanges();
+                Console.WriteLine("");
+                Console.WriteLine("SE HA ELIMINADO LA CATEGORIA");
+                return categoriaAEliminar;
+            }
+            else
+            {
+                Console.WriteLine("El número de ID ingresado no es válido.");
+                return categoriaAEliminar;
+            }
         }
     }
 }
